Fix valueToPercent to cover correct range
I was allowing the percent to be computed using a weird scaling that was
overshooting the maximum of my range. This was because I subtracted the minimum
range from the given value, but I didn't do that for the max range. The result
was basically lowering the value by the minimum range but allowing for values
over the maximum to still be computed.