Implement ultrasonics more generally
This is the first move in an expansion to simplify sensor interactions. Sensors
are now a class all to themselves, which will later be expanded to include
optical sensors. Ultrasonics are also implemented as an abstract class. Among
ultrasonics, there are different types. The first type is the one used in the
UltraPilot Mark I - the UltrasonicGeneralRanger, which reports distance over a
linearly scaled alert level.

Next, a new Ultrasonic will be implemented that polls for variability in
observations. This will be used to identify bad footing and obstacles in front
of the user's feet.